+++ src/utils/object/tabData.jsx
@@ -1,0 +1,26 @@
+import {
+	ClothesTop,
+	ClothesBottom,
+	Shoes,
+	Cap,
+} from '@icons/objecticon/objectIcon.js';
+
+// 탭 데이터
+const tabs = [
+	{
+		key: 'clothesTop',
+		title: '상의',
+		subtitle: 'clothes top',
+		icon: ClothesTop,
+	},
+	{
+		key: 'clothesBottom',
+		title: '하의',
+		subtitle: 'clothes bottom',
+		icon: ClothesBottom,
+	},
+	{ key: 'shoes', title: '신발', subtitle: 'shoes', icon: Shoes },
+	{ key: 'caps', title: '모자', subtitle: 'caps', icon: Cap },
+];
+
+export default tabs;
